Add check of emptiness when one field is left after removal
When there was two fields and then one of then was removed,
and it was the empty one, that disable the submit button, the
button was not enabled again.
A new function is added to check if the remaining fields are empty
or not.
The button state is updated accordingly.